perf(lysis): widen NodeId/EffectId/SsaVar to 64-bit indices
The lysis emission counter exhausts at u32::MAX = 4.29B emissions, which
ECDSAVerify-class circuits hit during execute. Widen the three coupled
index types end-to-end so multi-billion-emission programs no longer
panic on `node index + 1 overflows u32`.

NodeId: NonZeroU32 -> NonZeroU64 (lysis-types). EffectId: NonZeroU32 ->
NonZeroU64 (lysis intern). SsaVar: u32 -> u64 (ir-core newtype) with
IrProgram::next_var and getters/setters widening accordingly. The
lysis-to-R1CS bridge cast widens to `as u64` so the bridge no longer
truncates on >=2^32 ids. Cascade through every site that names
NodeId.raw(), SsaVar.0, or the bridge cast (ir/, ir-forge/lysis_lift,
instantiate, zkc).

Pinned by fresh_node_id_advances_past_u32_max in lysis-types/src/node.rs.
The Option<NodeId> niche stays single-word post-widening.

+++ ir-forge/src/instantiate/sink.rs
@@ -72,7 +72,7 @@ pub(crate) trait InstrSink<F: FieldBackend> {
     /// Current SSA var watermark (the id the next [`fresh_var`] would
     /// return). Used by `set_next_var` callers and by the
     /// canonicaliser.
-    fn next_var(&self) -> u32;
+    fn next_var(&self) -> u64;
 
     /// Begin a symbolic loop body — switch the sink's active push
     /// target to a fresh sub-buffer that
@@ -221,7 +221,7 @@ impl<'a, F: FieldBackend> InstrSink<F> for ExtendedSink<'a, F> {
         self.metadata.set_input_span(name, span);
     }
 
-    fn next_var(&self) -> u32 {
+    fn next_var(&self) -> u64 {
         self.metadata.next_var()
     }
